+++ src/ais/index.ts
@@ -7,6 +7,7 @@ import { HomeworldsAI } from "./homeworlds";
 import { ChaseAI } from "./chase";
 import { AbandeAI } from "./abande";
 import { CephalopodAI } from "./ceph";
+import { LinesOfActionAI } from "./loa";
 
 export interface IAIResult {
     bestMove: string|null;
@@ -17,7 +18,7 @@ export interface IAI {
     findmove: (state: any, depth: number) => string;
 }
 
-export { AIBase, AmazonsAI, BlamAI, CannonAI, MchessAI, HomeworldsAI, ChaseAI, AbandeAI, CephalopodAI };
+export { AIBase, AmazonsAI, BlamAI, CannonAI, MchessAI, HomeworldsAI, ChaseAI, AbandeAI, CephalopodAI, LinesOfActionAI };
 
 export const supportedGames: string[] = ["amazons"];
 export const fastGames: Map<string, number> = new Map([
@@ -29,6 +30,7 @@ export const fastGames: Map<string, number> = new Map([
     ["chase", 1],
     ["abande", 2],
     ["ceph", 4],
+    ["loa", 4],
 ]);
 export const slowGames: Map<string, number> = new Map([
     ["amazons", 2],
@@ -39,6 +41,7 @@ export const slowGames: Map<string, number> = new Map([
     ["chase", 2],
     ["abande", 3],
     ["ceph", 5],
+    ["loa", 5],
 ]);
 
 export function AIFactory(game: string): AIBase|undefined {
@@ -58,7 +61,9 @@ export function AIFactory(game: string): AIBase|undefined {
         case "abande":
             return new AbandeAI();
         case "ceph":
-            return new CephalopodAI()
+            return new CephalopodAI();
+        case "loa":
+            return new LinesOfActionAI();
     }
     return;
 }